scripts/requirements.txt
soundfile 
librosa
torch
openai-whisper
speechbox
transformers 
pyannote.audio==2.1.1
contextlib2
joblib
faster-whisper
evaluate
namedtupled
ctranslate2
feedparser
langchain
openai
tiktoken 
qdrant-client
nltk
python-dotenv
protobuf==3.20.3
funcy